refactor: move components to folders
