+++ src/components/Abutton/index.js
@@ -1,0 +1,3 @@
+import AButton from './AButton.vue';
+
+export default AButton;
