+++ src/components/AnimatedNumber/AnimatedNumber.vue
