+++ src/components/AnimatedNumber/index.js
@@ -1,0 +1,3 @@
+import AnimatedNumber from './AnimatedNumber.vue';
+
+export default AnimatedNumber;
