+++ src/components/FormElements/AForm/index.js
@@ -1,0 +1,3 @@
+import AForm from './AForm.vue';
+
+export default AForm;
