+++ src/components/FormElements/AInput/index.js
@@ -1,0 +1,3 @@
+import AInput from './AInput.vue';
+
+export default AInput;
